moved previous client to deprecated, added files for new client
