+++ google/cloud/bigtable/deprecated/__init__.py
@@ -1,5 +1,4 @@
-# -*- coding: utf-8 -*-
-# Copyright 2023 Google LLC
+# Copyright 2015 Google LLC
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,11 +11,15 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-#
+
+"""Google Cloud Bigtable API package."""
+
+from google.cloud.bigtable.client import Client
+
 from google.cloud.bigtable import gapic_version as package_version
+
+__version__: str
 
 __version__ = package_version.__version__
 
-from google.cloud.bigtable_async.client import BigtableDataClient
-
-__all__ = ("BigtableDataClient")
+__all__ = ["__version__", "Client"]
