+++ google/cloud/bigtable/deprecated/batcher.py
