+++ google/cloud/bigtable/deprecated/column_family.py
