+++ google/cloud/bigtable/deprecated/enums.py
