+++ google/cloud/bigtable/deprecated/instance.py
