+++ google/cloud/bigtable/deprecated/row.py
