+++ google/cloud/bigtable/deprecated/table.py
